pyproject.toml
[tool.poetry]
name = "telegrambot"
version = "0.1.0"
description = ""
authors = ["Oromion <caznaranl@uni.pe>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^12.8"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
